+++ micro-ui/web/micro-ui-internals/packages/libraries/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.9.2] [13-November-2025]
+### Bug fix
+ - Fixed the logout redirection issue
+ - added back the support for LOGOUT_REDIRECT_URL config 
+
+
+## [1.9.1] [12-November-2025]
+### Bug fix
+ - Fixed the logout redirection issue
+
 ## [1.9.0] [28-October-2025]
 
 ### 🚀 Enhanced Libraries & Core Utilities